requirements.txt
Django
django-taggit
wagtail
